Social/newsCrawler/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# news_process.py: 1
# newsspider.py: 1
# scrapy_selenium/http.py: 3
Scrapy == 1.5.0

# newsspider.py: 8
beautifulsoup4 == 4.6.0

# comment_classifier.py: 1
nltk == 3.3

# comment_classifier.py: 2
numpy == 1.14.3

# process_data.py: 9
pandas == 0.22.0

# news_crawler.py: 2
python_dateutil == 2.7.3

# news_auth.py: 2
python_dotenv == 0.8.2

# news_auth.py: 3
# news_crawler.py: 5
requests == 2.18.1

# comment_classifier.py: 3
scikit_learn == 0.19.1

